pyproject.toml
[tool.poetry]
name = "display-image"
version = "0.1.0"
description = ""
authors = ["edayot <pro.e.dayot@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.2.0"
rich = "^13.7.1"
click = "^8.1.7"
requests = "^2.31.0"
pokebase = "^1.4.1"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.poetry.scripts]
pokeinfo = "display_image:main"
